Fixed the aspect ratio of the ASCII art (aproximately)
In most fonts the height of a character is roughly twice the width of the character. Taking this into account meant that the number of outputs in the verticle dimmension had to be half that of the horizontal dimmension.

Changed dropout after conv layers to 0.2.

Added dropout of 0.1 after the input layer.

Added another 128 filter conv layer before the final output layer.