Added a data generator to lazily load data from a given directory during training
This means that there is now no delay when the script is run to load the data into RAM and process it prior to training. This also means that all of the training data can now be used since there is no limitation on how much can be fit in memory.